added user

+++ pages/TaskIndex.jsx
@@ -4,6 +4,8 @@ import { TaskFilter } from "../cmps/TaskFilter.jsx"
 import { TaskList } from "../cmps/TaskList.jsx"
 import { taskService } from "../services/task.service.js"
 import { showSuccessMsg } from "../services/event-bus.service.js"
+import { userService } from '../services/user.service.js'
+
 
 const { useState, useEffect } = React
 
@@ -11,6 +13,8 @@ export function TaskIndex() {
 
     const [tasks, setTasks] = useState(null)
     const [filterBy, setFilterBy] = useState(taskService.getDefaultFilter())
+    const user =userService.getLoggedinUser() || null
+
 
     useEffect(() => {
         loadTasks()
@@ -24,16 +28,35 @@ export function TaskIndex() {
     }
 
     function onRemoveTask(taskId) {
+        if(!user) return
         taskService.remove(taskId)
             .then(() => {
-              
-                setTasks(prevTasks => {
+                userService.addActivity('remove',user)
+                .then(()=> { setTasks(prevTasks => {
                     return prevTasks.filter(task => task.id !== taskId)
                 })
-                showSuccessMsg(`Task successfully removed!`)
+                showSuccessMsg(`Task successfully removed!`)})
+               
             })
             .catch(err => console.log('err:', err))
 
+    }
+
+    function onToggleStatus(task){
+        task.isActive=!task.isActive
+        taskService.save(task)
+        .then(loadTasks)
+    }
+
+    function onEditTask(taskId=0){
+        const txt=prompt('what is the task?')
+        if(!txt) return
+        taskService.save({txt,id:taskId})
+        .then(()=>{
+            if(taskId)        userService.addActivity('update',user)
+            if(!taskId)        userService.addActivity('add',user)
+        })
+        .then(loadTasks)
     }
 
 
@@ -43,14 +66,15 @@ export function TaskIndex() {
     }
 
     const { txt , status } = filterBy
-
     if (!tasks) return <div>Loading...</div>
     return (
-        <section className="task-index ">
+        <section style={user?{backgroundColor:user.prefs.bgcolor,color:user.prefs.color}:{}} className="task-index ">
             <h1>What is there to do?</h1>
             <TaskFilter filterBy={{ txt, status }} onSetFilter={onSetFilter} />
-            <Link to="/task/edit">Add</Link>
-            <TaskList tasks={tasks} onRemoveTask={onRemoveTask} />
+            {user && <button onClick={() => onEditTask()}>Add</button>}
+            {!tasks.length && <div>There are no tasks yet</div>}
+
+            <TaskList tasks={tasks} onRemoveTask={onRemoveTask} onEditTask={onEditTask} onToggleStatus={onToggleStatus} user={user} />
         </section>
     )
 }
